+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* **New** - Elementor Dynamic Tag: Resaweb URL No Accommodation
+
 ### 1.1.6: September 11th, 2018
 * **New** - Order package/accommodation/spatreatment/discovery by order_menu
 * **New** - Elementor Dynamic Tag: No Accommodation / Package Price